database/requirements.txt
pymongo==4.6.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
